homework_1/P1/requirements.txt
numpy
pandas
networkx
matplotlib.pyplot